+++ apps/backend/src/admin/admin.controller.ts
@@ -6,33 +6,22 @@ import {
   Body,
   Param,
   UseGuards,
-  UseInterceptors,
-  UploadedFile,
   Query,
   Headers,
 } from '@nestjs/common';
 import { AdminService } from './admin.service';
 import { Admin } from '@prisma/client';
-import { FileInterceptor } from '@nestjs/platform-express';
-import * as fs from 'fs';
-import { Multer } from 'multer';
+
 import { JwtAuthGuard } from '@auth';
-import { CreateAddNewsDto, CreateAdminDto, UpdateNewsDto } from './dto';
+import { CreateAdminDto } from './dto';
 import {
   CreateUserDto,
   ResponseCreateUser,
-  ResponseDeleteNews,
   ResponseDeleteUser,
   ResponseGetAllusers,
   ResponseLogin,
   ResponseLogout,
-  ResponseMessageEmail,
-  ResponseTips,
-  ResponseTipsDetails,
-  ResponseTipNews,
   ResponseToggleSubscription,
-  ResponseAddNews,
-  ResponseUpdateNews,
 } from '@shared/dtos';
 
 @Controller('admin')
@@ -96,97 +85,5 @@ export class AdminController {
     @Param('id') id: string
   ): Promise<ResponseToggleSubscription> {
     return await this.adminService.toggleSubscription(parseInt(id, 10));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post('tips-add')
-  @UseInterceptors(FileInterceptor('file'))
-  async addTips(
-    @Body() tipsData: { title: string; description: string },
-    @UploadedFile() file?: Express.Multer.File | undefined
-  ) {
-    let htmlContent;
-    if (file) {
-      htmlContent = fs.readFileSync(file.path, 'utf8'); // Use file.path directly
-    }
-    return await this.adminService.addTips(tipsData, htmlContent);
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('tips')
-  async getTips(): Promise<ResponseTips> {
-    return await this.adminService.getTips();
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('tips/:id')
-  async getTipsById(
-    @Param('id') id: string,
-    @Query('page') page?: string,
-    @Query('limit') limit?: string
-  ): Promise<ResponseTipsDetails> {
-    const pageNumber = parseInt(page, 10);
-    const limitNumber = parseInt(limit, 10);
-    return await this.adminService.getTipsById(
-      parseInt(id, 10),
-      pageNumber,
-      limitNumber
-    );
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post('news/add')
-  @UseInterceptors(FileInterceptor('file'))
-  async addNews(
-    @UploadedFile() file: Express.Multer.File,
-    @Body() newsData: CreateAddNewsDto
-  ): Promise<ResponseAddNews> {
-    if (!file) {
-      console.error('File not provided');
-      throw new Error('File not provided');
-    }
-
-    const htmlContent = fs.readFileSync(file.path, 'utf8'); // Use file.path directly
-
-    return await this.adminService.addNews(newsData, htmlContent);
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('tips/news/:id')
-  async getNews(
-    @Param('id') id: string,
-    @Query('page') page?: string,
-    @Query('limit') limit?: string
-  ): Promise<ResponseTipNews> {
-    const pageNumber = parseInt(page, 10);
-    const limitNumber = parseInt(limit, 10);
-    return await this.adminService.getNews(
-      parseInt(id, 10),
-      pageNumber,
-      limitNumber
-    );
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('news/:id')
-  async getNewsById(@Param('id') id: string) {
-    return await this.adminService.getNewsById(parseInt(id, 10));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @UseInterceptors(FileInterceptor('file')) // <-- Add this line
-  @Post('news/update/:id')
-  async updateNews(
-    @Param('id') id: string,
-    @Body() newsData: UpdateNewsDto,
-    @UploadedFile() file: Express.Multer.File
-  ): Promise<ResponseUpdateNews> {
-    return await this.adminService.updateNews(parseInt(id, 10), newsData, file);
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Delete('news/:id')
-  async deleteNews(@Param('id') id: string): Promise<ResponseDeleteNews> {
-    return await this.adminService.deleteNews(parseInt(id, 10));
   }
 }
