+++ apps/backend/src/admin/admin.module.ts
@@ -11,9 +11,7 @@ import {
 import { ConfigModule } from '@nestjs/config';
 import generalConfig from '@shared/config/general.config';
 import authConfig from '@auth/config/auth.config';
-import { AuthModule, AuthService } from '@auth';
-import { LogsDatabaseService } from '@database/logs';
-import { WinstonLoggerService } from '@logger-winston';
+import { AuthModule } from '@auth';
 
 @Module({
   imports: [
