+++ apps/backend/src/admin/dto/admin.dto.ts
@@ -47,27 +47,3 @@ export class SendEmailDto {
   @IsNotEmpty()
   html: string;
 }
-
-export class CreateAddNewsDto {
-  @IsNotEmpty()
-  @IsString()
-  title: string;
-
-  @IsNotEmpty()
-  @IsString()
-  tipsId: string;
-
-  @IsOptional()
-  @IsString()
-  status: 'true' | 'false';
-}
-
-export class UpdateNewsDto {
-  @IsOptional()
-  @IsString()
-  title?: string;
-
-  @IsOptional()
-  @IsString()
-  status?: 'true' | 'false';
-}
