+++ apps/backend/src/news/dtos/index.ts
@@ -1,0 +1,1 @@
+export * from './news.dtos';
