+++ apps/frontend/src/app/auto-news/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Navbar } from '../components';
-import { CronTime } from '../components/auto-email/CronTime';
+import { CronTime } from '../components/auto-news/CronTime';
 const CronPage = () => {
   return (
     <>
