+++ apps/frontend/src/app/components/auto-news/cronName.tsx
