+++ apps/frontend/src/app/components/auto-news/index.ts
