+++ apps/frontend/src/app/components/auto-news/startTimer.tsx
