--- apps/frontend/src/app/components/emails/index.ts
@@ -1,1 +1,0 @@
-export * from './emails';
