+++ apps/frontend/src/app/components/index.ts
@@ -1,6 +1,6 @@
-export * from './auto-email';
+export * from './auto-news';
 export * from './dtos';
-export * from './email-order';
-export * from './emails';
+export * from './news-order';
+export * from './news';
 export * from './navbar';
 export * from './tips';
