+++ apps/frontend/src/app/components/news-order/email-order.tsx
@@ -1,11 +1,11 @@
 'use client';
-import { ResponseEmailOrderDto } from '@shared/dtos';
+import { ResponseNewsOrderDto } from '@shared/dtos';
 import { fetchWithAuth } from '@utils';
 import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 
 export const Emailorder = () => {
-  const [data, setData] = useState<ResponseEmailOrderDto['data'] | null>(null);
+  const [data, setData] = useState<ResponseNewsOrderDto['data'] | null>(null);
   const router = useRouter();
 
   useEffect(() => {
@@ -15,7 +15,7 @@ export const Emailorder = () => {
     }
     const fetchEmailOrder = async () => {
       try {
-        const res: ResponseEmailOrderDto = await fetchWithAuth(
+        const res: ResponseNewsOrderDto = await fetchWithAuth(
           'email/emailsorder',
           {},
           true
