+++ apps/frontend/src/app/news/page.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { Emails, Navbar } from '../components';
+import { AllEmails, Navbar } from '../components';
 
 const EmailsPage: React.FC = () => {
   return (
     <div>
       <Navbar />
-      <Emails />
+      <AllEmails />
     </div>
   );
 };
