+++ apps/frontend/src/app/tips/page.tsx
@@ -32,7 +32,7 @@ const TipsPage: React.FC = () => {
     }
     const fetchAllTips = async () => {
       try {
-        const res: ResponseTips = await fetchWithAuth(`admin/tips`, {}, true); // Fetch all users
+        const res: ResponseTips = await fetchWithAuth(`news/tips`, {}, true); // Fetch all users
 
         const data = res.data;
         if (data && Array.isArray(data.tips)) {
